chore: refactor mandatee-selector component
