+++ app/components/mandatees/mandatees-selector-modal.hbs
@@ -23,7 +23,6 @@
     >
       <Auk::Label>{{t "minister"}}</Auk::Label>
       <Mandatees::MandateeSelector
-        @openSearch={{this.openSearch}}
         @excludeMandatees={{@excludeMandatees}}
         @referenceDate={{@referenceDate}}
         @openSearch={{this.openSearch}}
