+++ translations/nl-be.json
@@ -1230,5 +1230,6 @@
   "sign-flow-was-sent-while-you-were-editing-could-not-add-new-version": "Dit document is naar de handtekenmap verzonden. Het is niet mogelijk om een nieuwe versie op te laden.",
   "sign-flow-was-sent-while-you-were-editing-could-not-edit": "Dit document is naar de handtekenmap verzonden. Het is niet mogelijk om de gegevens van het document aan te passen.",
   "sign-flow-was-sent-while-you-were-editing-could-not-delete": "Dit document is naar de handtekenmap verzonden. Het is niet mogelijk om het document in Kaleidos te verwijderen.",
-  "sign-flow-was-sent-cannot-stop-it": "De ondertekenflow is opgestart. Het is niet mogelijk om de flow te stoppen. Neem contact op met de Secretarie."
+  "sign-flow-was-sent-cannot-stop-it": "De ondertekenflow is opgestart. Het is niet mogelijk om de flow te stoppen. Neem contact op met de Secretarie.",
+  "type-to-search": "Typ om te zoeken"
 }
